+++ script/run.sh
@@ -15,7 +15,7 @@
 
 set -e
 
-VERSION="1.6.0"
+VERSION="1.6.1"
 IMAGE="docker/compose:$VERSION"
 
 
@@ -31,7 +31,9 @@ fi
 
 
 # Setup volume mounts for compose config and context
-VOLUMES="-v $(pwd):$(pwd)"
+if [ "$(pwd)" != '/' ]; then
+    VOLUMES="-v $(pwd):$(pwd)"
+fi
 if [ -n "$COMPOSE_FILE" ]; then
     compose_dir=$(dirname $COMPOSE_FILE)
 fi
@@ -45,9 +47,10 @@ fi
 
 # Only allocate tty if we detect one
 if [ -t 1 ]; then
-    DOCKER_RUN_OPTIONS="-ti"
-else
-    DOCKER_RUN_OPTIONS="-i"
+    DOCKER_RUN_OPTIONS="-t"
+fi
+if [ -t 0 ]; then
+    DOCKER_RUN_OPTIONS="$DOCKER_RUN_OPTIONS -i"
 fi
 
-exec docker run --rm $DOCKER_RUN_OPTIONS $DOCKER_ADDR $COMPOSE_OPTIONS $VOLUMES -w $(pwd) $IMAGE $@
+exec docker run --rm $DOCKER_RUN_OPTIONS $DOCKER_ADDR $COMPOSE_OPTIONS $VOLUMES -w "$(pwd)" $IMAGE "$@"